router/requirements.txt
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
